Adjusted codebase
Adjusted codebase to work on local server, although data folder containing social media data is not provided for the purpose of protecting the organization. Functions used to retrieve social media from Google Drive is commented out. Service Account json file does not contain the private key to again for the purpose of the organization's security.